requirements.txt
kivy==2.2.1
numpy==1.24.3
opencv-python==4.8.1.78
pytesseract==0.3.10
pyautogui==0.9.54
pillow==10.0.0
mss==9.0.1
pywin32==306
psutil==5.9.5 